+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - Add support for JAX pmap ([GH-976](https://github.com/NVIDIA/warp/pull/976)).
 - Add adjoint for `wp.transform()` when constructing with individual scalars ([GH-1011](https://github.com/NVIDIA/warp/issues/1011)).
 - Add a double precision overload for `wp.intersect_tri_tri` ([GH-1015](https://github.com/NVIDIA/warp/issues/1015)).
+- Add `wp.get_cuda_supported_archs()` to query supported CUDA compute architectures for compilation targets
+  ([GH-964](https://github.com/NVIDIA/warp/issues/964)).
 
 ### Removed